Extracting Theme variable, created new file Theme.js, exporting Theme object
