+++ sauti/src/Components/Chart/AgeChart.js
@@ -1,45 +1,9 @@
 import React from "react";
 import { ResponsiveBar } from "@nivo/bar";
 import axios from "axios";
+import theme from '../../Constants/Theme.js'
 
-const theme = {
-  //background: "#222222",
-  axis: {
-    // fontSize: "100px",
-    // tickColor: "#eee",
-    ticks: {
-      // line: {
-      //   stroke: "#555555"
-      // },
-      text: {
-        fill: "#4d4f52",
-        fontSize: "1.6rem"
-      }
-    },
-    legend: {
-      text: {
-        fill: "#4d4f52",
-        fontSize: "1.6rem",
-        fontWeight: 550,
-        fontFamily: "Helvetica"
-      }
-    }
-  },
-  labels: {
-    text: {
-      fontSize: "1.6rem",
-      fontWeight: 550,
-      fontFamily: "Helvetica",
-      fill: "#4d4f52",
-    }
-  }
-    // Grid Lines
-  // grid: {
-  //   line: {
-  //     stroke: "#555555"
-  //   }
-  // }
-};
+
 
 class AgeChart extends React.Component {
   constructor(props) {
